Feat: Allow bucket region to be specified in s3auth.conf or $AWS_REGION
The `/etc/apt/s3auth.conf` may contain the AWS bucket region and the
AWS IAM credentials that will be used to pull from the bucket.

However, if the file does not contain credentials (in particular `AccessKeyId`
then the region will be pulled from `$AWS_REGION` and lastly from the
ec2 meta-data service.

This change allows the region to be set in `/etc/apt/s3auth.conf`
even if the file does not contain any credentials. They will be loaded
from the ec2 meta-data service.

This allows an instance (using an instance role) to pull from a bucket
in a different region without needing to set $AWS_REGION for apt (which
may influence / alter how installed packages function)