Feat: Support following 301 redirects when using incorrect bucket region
When making a request to an S3 endpoint for a bucket in a different region
S3 will return a 301 redirect and include the correct region in
the `x-amz-bucket-region` response header.

The default region is specified in the config:

  1. In /etc/apt/s3auth.conf with Region = "region-name"
  2. In $AWS_REGION
  3. In the ec2 meta-data

This will be used when creating the endpoint hostname.

If the request to this endpoint 301 + x-amz-bucket-region, then
the request will repeat but using a new endpoint name and the
new region in the AWSv4 signature.

The region is cached for the bucket name. This allows different
buckets in different regions to be used.

If the regional requests continue to 301, we'll give up after 5
attempts (RETRIES = 5)

If an endpoint hostname is set (in s3auth.conf), this will
inhibit the region redirection.